+++ ecosistema_ia/agentes/tipos/sublimes/calificador.py
@@ -1,22 +1,25 @@
-# calificador.py
+"""Agent that evaluates other agents' performance."""
 
-class Calificador:
-    def __init__(self, ciclos_observacion=3):
-        self.observaciones = {}  # {id_agente: [excreciones]}
+from agentes.tipos.sublimes.sublime_base import SublimeBase
+
+
+class Calificador(SublimeBase):
+    def __init__(self, identificador="CAL-001", x=0, y=0, z=0, ciclos_observacion=3):
+        super().__init__(identificador, x, y, z, funcion="calificador")
         self.ciclos_observacion = ciclos_observacion
+        self.observaciones = {}  # {id_agente: [excreciones]}
 
-    def observar(self, agentes):
+    def observar(self, territorio, agentes, ciclo):
         for agente in agentes:
             if agente.identificador not in self.observaciones:
                 self.observaciones[agente.identificador] = []
 
-            # Registrar excreción actual
             self.observaciones[agente.identificador].append(agente.excrecion)
 
             # Mantener solo los últimos N ciclos
             self.observaciones[agente.identificador] = self.observaciones[agente.identificador][-self.ciclos_observacion:]
 
-            # Evaluar desempeño si hay suficientes ciclos
+            # Evaluar desempeño si hay suficientes datos
             if len(self.observaciones[agente.identificador]) == self.ciclos_observacion:
                 excreciones = self.observaciones[agente.identificador]
                 if any(e for e in excreciones if e is not None and str(e).strip()):
@@ -24,4 +27,6 @@ class Calificador:
                 else:
                     agente.calificacion = "roja"
 
-                print(f"🏁 {agente.identificador} calificado como: {agente.calificacion}")
+                print(f"🏁 {self.identificador} calificó a {agente.identificador} como: {agente.calificacion}")
+
+__all__ = ["Calificador"]
