+++ ecosistema_ia/agentes/tipos/sublimes/sublime_base.py
@@ -1,31 +1,14 @@
-# calificador.py
+"""Base class for sublime agents."""
 
-from agentes.tipos.sublimes.sublime_base import SublimeBase
+from agentes.agente_base import AgenteBase
 
-class Calificador(SublimeBase):
-    def __init__(self, identificador="CAL-001", x=0, y=0, z=0, ciclos_observacion=3):
-        super().__init__(identificador, x, y, z, funcion="calificador")
-        self.ciclos_observacion = ciclos_observacion
-        self.observaciones = {}  # {id_agente: [excreciones]}
 
-    def observar(self, territorio, agentes, ciclo):
-        for agente in agentes:
-            if agente.identificador not in self.observaciones:
-                self.observaciones[agente.identificador] = []
+class SublimeBase(AgenteBase):
+    """Minimal base class for high-level "sublime" agents."""
 
-            self.observaciones[agente.identificador].append(agente.excrecion)
+    def __init__(self, identificador, x, y, z, funcion="sublime"):
+        super().__init__(identificador, x, y, z, funcion=funcion)
 
-            # Mantener solo los últimos N ciclos
-            self.observaciones[agente.identificador] = self.observaciones[agente.identificador][-self.ciclos_observacion:]
 
-            # Evaluar desempeño si hay suficientes datos
-            if len(self.observaciones[agente.identificador]) == self.ciclos_observacion:
-                excreciones = self.observaciones[agente.identificador]
-                if any(e for e in excreciones if e is not None and str(e).strip()):
-                    agente.calificacion = "verde"
-                else:
-                    agente.calificacion = "roja"
+__all__ = ["SublimeBase"]
 
-                print(f"🏁 {self.identificador} calificó a {agente.identificador} como: {agente.calificacion}")
-
-__all__ = ["Calificador"]
